fixed sanity error
