+++ next.config.mjs
@@ -1,7 +1,26 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
     images : {
-        domains : ['cdn.sanity.io']
+        remotePatterns : [
+            {
+                protocol : 'https',
+                hostname : 'lh3.googleusercontent.com',
+                port : '',
+                pathname : '/**'
+            },
+            {
+                protocol : 'https',
+                hostname : 'cdn.sanity.io',
+                port : '',
+                pathname : '/**'
+            },
+            {
+                protocol : 'https',
+                hostname : 'images.unsplash.com',
+                port : '',
+                pathname : '/**'
+            }
+        ],
     },
     reactStrictMode : false
 };
