+++ src/app/api/follow/route.ts
@@ -1,7 +1,5 @@
-import { getServerSession } from "next-auth";
+
 import { NextRequest, NextResponse } from "next/server";
-import { authOptions } from "../auth/[...nextauth]/route";
-import { dislikePost, likePost } from "@/app/service/posts";
 import { follow, unfollow } from "@/app/service/user";
 import { withSessionUser } from "@/app/util/session";
