+++ src/app/auth/signin/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { authOptions } from '../../api/auth/[...nextauth]/route'
+import { authOptions } from "@/app/auth/authOptions";
 import { getServerSession } from 'next-auth'
 import { redirect } from 'next/navigation';
 import { getProviders } from 'next-auth/react';
