+++ src/app/page.tsx
@@ -1,9 +1,9 @@
-import Image from "next/image";
+
 import FollowingBar from "./components/FollowingBar";
 import PostList from "./components/PostList";
 import SideBar from "./components/SideBar";
 import { getServerSession } from "next-auth";
-import { authOptions } from "./api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/auth/authOptions";
 import { redirect } from "next/navigation";
 import { signIn } from "next-auth/react";
